requirements.txt
django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
coreapi==2.3.3